# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for payment sandbox simulation.",
        "title": "Payment Sandbox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/payment-intents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin lists payment intents with optional status filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "List payment intents",
                "parameters": [
                    {
                        "enum": [
                            "PENDING",
                            "SUCCESS",
                            "FAILED"
                        ],
                        "type": "string",
                        "description": "Payment intent status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/app_modules_payment_handlers.PaymentIntentWithAmount"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/payment-intents/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin updates payment intent status to SUCCESS or FAILED",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Update payment intent status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment intent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment status payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_modules_payment_handlers.UpdatePaymentIntentStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_payment_handlers.PaymentIntentStatusUpdateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/pay/{token}": {
            "get": {
                "description": "Public endpoint to fetch invoice detail by payment link token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Get invoice by payment token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_payment_handlers.PublicInvoiceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pay/{token}/intents": {
            "post": {
                "description": "Public endpoint to create payment intent for an invoice token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Create payment intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment intent payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_modules_payment_handlers.CreatePaymentIntentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_payment_handlers.PaymentIntentCreateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_modules_payment_handlers.CreatePaymentIntentRequest": {
            "type": "object",
            "required": [
                "method"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "WALLET",
                        "VA_DUMMY",
                        "EWALLET_DUMMY"
                    ],
                    "example": "WALLET"
                }
            }
        },
        "app_modules_payment_handlers.PaymentIntentCreateResponse": {
            "type": "object",
            "properties": {
                "invoice": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.Invoice"
                },
                "payment_intent": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentIntent"
                }
            }
        },
        "app_modules_payment_handlers.PaymentIntentStatusUpdateResponse": {
            "type": "object",
            "properties": {
                "invoice": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.Invoice"
                },
                "payment_intent": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentIntent"
                }
            }
        },
        "app_modules_payment_handlers.PaymentIntentWithAmount": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentMethod"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "app_modules_payment_handlers.PublicInvoiceResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "payment_link_token": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "app_modules_payment_handlers.UpdatePaymentIntentStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "SUCCESS",
                        "FAILED"
                    ],
                    "example": "SUCCESS"
                }
            }
        },
...
        "payment-sandbox_app_modules_invoice_models_entity.Invoice": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "payment_link_token": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "PAID",
                "EXPIRED"
            ],
            "x-enum-varnames": [
                "InvoicePending",
                "InvoicePaid",
                "InvoiceExpired"
            ]
        },
...
        "payment-sandbox_app_modules_payment_models_entity.PaymentIntent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentMethod"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_payment_models_entity.PaymentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "payment-sandbox_app_modules_payment_models_entity.PaymentMethod": {
            "type": "string",
            "enum": [
                "WALLET",
                "VA_DUMMY",
                "EWALLET_DUMMY"
            ],
            "x-enum-varnames": [
                "MethodWallet",
                "MethodVADummy",
                "MethodEWalletDummy"
            ]
        },
        "payment-sandbox_app_modules_payment_models_entity.PaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentSuccess",
                "PaymentFailed"
            ]
        },
...
        "payment-sandbox_app_shared_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                },
                "meta": {}
            }
        },
        "payment-sandbox_app_shared_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}